refactor: extract fetch interceptor from main.tsx to DirectModeInterceptor
Move ~1350 lines of direct-mode fetch interception logic from main.tsx
to services/interceptor/DirectModeInterceptor.ts. main.tsx now calls
installDirectModeInterceptor() behind the isDirectMode() guard.

- Extract all route handlers, profile cache, and prefetch logic
- Move jsonResponse helper and CachedProfile interface
- Change catch-all from silent {} to 501 Not Implemented with console.warn
- Zero behavioral change for all handled routes
